+++ src/main/java/com/mindDiary/mindDiary/service/UserServiceImpl.java
@@ -15,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
@@ -29,6 +30,7 @@ public class UserServiceImpl implements UserService {
   private final RedisStrategy redisStrategy;
   private final EmailStrategy emailStrategy;
   private final TokenStrategy tokenStrategy;
+  private final UserTransactionService userTransactionService;
 
 
   @Override
@@ -45,21 +47,9 @@ public class UserServiceImpl implements UserService {
     redisStrategy.addEmailToken(user.getEmailCheckToken(), user.getId());
     emailStrategy.sendUserJoinMessage(user.getEmailCheckToken(), user.getEmail());
 
-    removeCacheAfterRollback(user.getEmailCheckToken());
+    userTransactionService.removeCacheAfterRollback(user.getEmailCheckToken());
   }
 
-
-  public void removeCacheAfterRollback(String emailToken) {
-    TransactionSynchronizationManager.registerSynchronization(
-        new TransactionSynchronization() {
-          @Override
-          public void afterCompletion(int status) {
-            if (status == STATUS_ROLLED_BACK) {
-              redisStrategy.deleteValue(emailToken);
-            }
-          }
-        });
-  }
 
 
   public void isNicknameDuplicate(String nickname) {
