+++ src/main/java/com/mindDiary/mindDiary/service/UserTransactionService.java
@@ -1,0 +1,5 @@
+package com.mindDiary.mindDiary.service;
+
+public interface UserTransactionService {
+  void removeCacheAfterRollback(String emailToken);
+}
